requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --index-url=https://pypi.org/simple requirements.in
#
aiosqlite==0.20.0
    # via prefect
alembic==1.14.0
    # via prefect
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   prefect
    #   starlette
apprise==1.9.0
    # via prefect
asgi-lifespan==2.1.0
    # via prefect
asyncpg==0.30.0
    # via prefect
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
cachetools==5.5.0
    # via prefect
certifi==2024.12.14
    # via
    #   apprise
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   apprise
    #   prefect
    #   typer
    #   uvicorn
cloudpickle==3.1.0
    # via prefect
colorama==0.4.6
    # via griffe
coolname==2.2.0
    # via prefect
croniter==5.0.1
    # via prefect
cryptography==44.0.0
    # via prefect
dateparser==1.2.0
    # via prefect
deprecated==1.2.15
    # via opentelemetry-api
docker==7.1.0
    # via prefect
exceptiongroup==1.2.2
    # via prefect
fastapi==0.115.6
    # via prefect
fsspec==2024.10.0
    # via prefect
graphviz==0.20.3
    # via prefect
greenlet==3.1.1
    # via sqlalchemy
griffe==1.5.1
    # via prefect
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.7
    # via
    #   httpx
    #   prefect
httpx[http2]==0.28.1
    # via prefect
humanize==4.11.0
    # via
    #   jinja2-humanize-extension
    #   prefect
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.5.0
    # via opentelemetry-api
jinja2==3.1.4
    # via
    #   jinja2-humanize-extension
    #   prefect
jinja2-humanize-extension==0.4.0
    # via prefect
jsonpatch==1.33
    # via prefect
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via prefect
jsonschema-specifications==2024.10.1
    # via jsonschema
mako==1.3.8
    # via alembic
markdown==3.7
    # via apprise
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.2.2
    # via requests-oauthlib
opentelemetry-api==1.29.0
    # via prefect
orjson==3.10.12
    # via prefect
packaging==24.2
    # via prefect
pathspec==0.12.1
    # via prefect
pendulum==3.0.0
    # via
    #   -r requirements.in
    #   prefect
prefect==3.1.4
    # via -r requirements.in
prometheus-client==0.21.0
    # via
    #   -r requirements.in
    #   prefect
pycparser==2.22
    # via cffi
pydantic==2.10.3
    # via
    #   fastapi
    #   prefect
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.27.1
    # via
    #   prefect
    #   pydantic
pydantic-extra-types==2.10.1
    # via prefect
pydantic-settings==2.7.0
    # via prefect
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   dateparser
    #   pendulum
    #   prefect
    #   time-machine
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
python-slugify==8.0.4
    # via prefect
pytz==2024.2
    # via
    #   croniter
    #   dateparser
    #   prefect
pyyaml==6.0.2
    # via
    #   apprise
    #   prefect
readchar==4.2.1
    # via prefect
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via dateparser
requests==2.32.3
    # via
    #   -r requirements.in
    #   apprise
    #   docker
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via apprise
rfc3339-validator==0.1.4
    # via prefect
rich==13.9.4
    # via
    #   prefect
    #   typer
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.6
    # via prefect
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   asgi-lifespan
    #   prefect
sqlalchemy[asyncio]==2.0.36
    # via
    #   alembic
    #   prefect
starlette==0.41.3
    # via fastapi
text-unidecode==1.3
    # via python-slugify
time-machine==2.16.0
    # via pendulum
toml==0.10.2
    # via prefect
typer==0.13.1
    # via prefect
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   alembic
    #   anyio
    #   fastapi
    #   prefect
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   sqlalchemy
    #   typer
tzdata==2024.2
    # via pendulum
tzlocal==5.2
    # via dateparser
ujson==5.10.0
    # via prefect
urllib3==2.2.3
    # via
    #   docker
    #   requests
uvicorn==0.34.0
    # via prefect
websockets==13.1
    # via prefect
wrapt==1.17.0
    # via deprecated
zipp==3.21.0
    # via importlib-metadata
    